move elasticDSL to utils
