+++ src/example-types/facet.js
@@ -2,7 +2,7 @@ let _ = require('lodash/fp')
 let F = require('futil')
 let { buildRegexQueryForWords } = require('../utils/regex')
 let { getField } = require('../fields')
-let { negate, elasticsearchIntegerMax } = require('../elasticDSL')
+let { negate, elasticsearchIntegerMax } = require('../utils/elasticDSL')
 
 module.exports = {
   hasValue: _.get('values.length'),
