+++ src/example-types/legacy/matchStats.js
@@ -1,6 +1,6 @@
 let _ = require('lodash/fp')
 let esTwoLevel = require('./esTwoLevelAggregation').result
-let { negate } = require('../../elasticDSL')
+let { negate } = require('../../utils/elasticDSL')
 
 module.exports = {
   validContext: node => node.key_field && node.value_field && node.key_value,
